pyproject.toml
[project]
name = "wildcamtools"
version = "0.0.1"
description = "Tools for manipulating wildlife camera clips"
authors = [
    { name = "Adam Faulconbridge", email = "afaulconbridge@googlemail.com" },
]
readme = "README.md"
keywords = ['python']
requires-python = "==3.13.*"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "ffmpeg-python>=0.2.0",
    "opencv-contrib-python-headless>=4.12.0.88",
    "psutil>=7.1.3",
    "typer>=0.20.0",
    "types-psutil>=7.0.0.20251111",
]
[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "mypy>=0.991",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
]

[project.urls]
Homepage = "https://afaulconbridge.github.io/wildcamtools/"
Repository = "https://github.com/afaulconbridge/wildcamtools"
Documentation = "https://afaulconbridge.github.io/wildcamtools/"


[project.scripts]
wildcamtools = "wildcamtools.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/wildcamtools"]

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py314"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.uv]
# we only want the headless version so put an impossible dependency on the gui version
# see https://github.com/astral-sh/uv/issues/14220
override-dependencies = ["opencv-python; python_version < '0'"]
